chore: Update dev.next.config.js.sample and next.config.js
This commit updates the dev.next.config.js.sample and next.config.js files. The dev.next.config.js.sample file is modified to export a config object, and the next.config.js file is updated to include the devconfig object. These changes are necessary to improve the configuration of the project.